:sparkles: Check for ES enabled

+++ src/crons/daily.ts
@@ -1,4 +1,4 @@
-import { elasticSearch } from "@staart/elasticsearch";
+import { elasticSearch, elasticSearchEnabled } from "@staart/elasticsearch";
 import { ms } from "@staart/text";
 import { CronJob } from "cron";
 import { ELASTIC_LOGS_INDEX } from "../config";
@@ -15,24 +15,25 @@ export default () => {
 };
 
 const deleteOldLogs = async () => {
-  return (
-    await elasticSearch.deleteByQuery({
-      index: ELASTIC_LOGS_INDEX,
-      body: {
-        query: {
-          bool: {
-            must: [
-              {
-                range: {
-                  date: {
-                    lte: new Date(new Date().getTime() - ms("92 days")),
+  if (elasticSearchEnabled)
+    return (
+      await elasticSearch.deleteByQuery({
+        index: ELASTIC_LOGS_INDEX,
+        body: {
+          query: {
+            bool: {
+              must: [
+                {
+                  range: {
+                    date: {
+                      lte: new Date(new Date().getTime() - ms("92 days")),
+                    },
                   },
                 },
-              },
-            ],
+              ],
+            },
           },
         },
-      },
-    })
-  ).body;
+      })
+    ).body;
 };
